backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Azure SDKs
azure-identity==1.19.0
azure-ai-projects==1.0.0b1
azure-keyvault-secrets==4.9.0
openai>=1.75.0
azure-search-documents==11.6.0
azure-storage-blob==12.23.0

# LangGraph & LangChain
langgraph==0.2.40
langchain==0.3.7
langchain-openai==0.2.9
langchain-core>=0.3.17

# Observability (simplified for local dev)
# opentelemetry-api==1.28.0
# opentelemetry-sdk==1.28.0
# opentelemetry-instrumentation-fastapi==0.57b0
# azure-monitor-opentelemetry==1.7.0

# Utils
httpx==0.27.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
redis==5.2.0

# Async support
aiofiles==24.1.0

# PostgreSQL & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
